handson/todo-app/requirements.txt
Flask
Flask-SQLAlchemyFlask-Script
flask-oidc